Update event.js
Hi,

I've been using your node module to connect to the Last.FM API, but I needed to do a search for festivals by name, which isn't described in the official documentation, but seems to exist nevertheless (at 'event.search', similar to 'artist.search'). Therefore, I've added this extra function to your module to be able to call this specific API function. I've pretty much copy-pasted it from artist.js, so not a lot of added magic here. Just thought it would be useful to also support this call.

Gr, Jiri